+++ client/src/adt/abap/AbapCds.ts
@@ -2,6 +2,6 @@ import { AbapObject } from "./AbapObject"
 
 export class AbapCds extends AbapObject {
   public getExtension(): string {
-    return ".cds"
+    return ".cds.abap"
   }
 }